refactor: Move out style code on container
관심사 분리를 위해 컨테이너에서 스타일 관련 코드를 분리하였습니다.
